Add build scripts; Modify run.bat to select user installed python or embedded python

+++ run.bat
@@ -1,6 +1,13 @@
 @setlocal enableextensions
 @cd /d "%~dp0"
 
-py -m pip install -r requirements.txt
-py src\app.py
+if exist .\Python39\python.exe (
+    echo "Use embedded python"
+    .\Python39\python.exe src\app.py
+) else (
+    echo "Use user python"
+    py -m pip install -r requirements.txt
+    py src\app.py
+)
+
 pause
